refactor: extract types to correct package
