+++ builder/builder/builder_test.go
@@ -492,17 +492,17 @@ func sseConstraintsHandler(w http.ResponseWriter, r *http.Request) {
 }
 
 // generateMockConstraintsForSlot generates a list of constraints for a given slot
-func generateMockConstraintsForSlot(slot uint64) common.SignedConstraintsList {
-	rawTx := new(common.HexBytes)
-	err := rawTx.UnmarshalJSON([]byte("\"0x02f876018305da308401312d0085041f1196d2825208940c598786c88883ff5e4f461750fad64d3fae54268804b7ec32d7a2000080c080a0086f02eacec72820be3b117e1edd5bd7ed8956964b28b2d903d2cba53dd13560a06d61ec9ccce6acb31bf21878b9a844e7fdac860c5b7d684f7eb5f38a5945357c\""))
+func generateMockConstraintsForSlot(slot uint64) types.SignedConstraintsList {
+	rawTx := new(types.Transaction)
+	err := rawTx.UnmarshalBinary(common.Hex2Bytes("0x02f876018305da308401312d0085041f1196d2825208940c598786c88883ff5e4f461750fad64d3fae54268804b7ec32d7a2000080c080a0086f02eacec72820be3b117e1edd5bd7ed8956964b28b2d903d2cba53dd13560a06d61ec9ccce6acb31bf21878b9a844e7fdac860c5b7d684f7eb5f38a5945357c"))
 	if err != nil {
 		fmt.Println("Failed to unmarshal rawTx: ", err)
 	}
 
-	return common.SignedConstraintsList{
-		&common.SignedConstraints{
-			Message: common.ConstraintsMessage{
-				Transactions: []*common.HexBytes{rawTx}, Pubkey: phase0.BLSPubKey{}, Slot: slot,
+	return types.SignedConstraintsList{
+		&types.SignedConstraints{
+			Message: types.ConstraintsMessage{
+				Transactions: []*types.Transaction{rawTx}, Pubkey: phase0.BLSPubKey{}, Slot: slot,
 			}, Signature: phase0.BLSSignature{},
 		},
 	}
