+++ builder/common/types.go
@@ -18,9 +18,7 @@ package common
 
 import (
 	"bytes"
-	"crypto/sha256"
 	"database/sql/driver"
-	"encoding/binary"
 	"encoding/hex"
 	"encoding/json"
 	"errors"
@@ -32,17 +30,7 @@ import (
 	"strings"
 
 	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/flashbots/go-boost-utils/bls"
 	"golang.org/x/crypto/sha3"
-
-	v1 "github.com/attestantio/go-builder-client/api/v1"
-	builderSpec "github.com/attestantio/go-builder-client/spec"
-	consensusSpec "github.com/attestantio/go-eth2-client/spec"
-	bellatrixSpec "github.com/attestantio/go-eth2-client/spec/bellatrix"
-	capellaSpec "github.com/attestantio/go-eth2-client/spec/capella"
-	denebSpec "github.com/attestantio/go-eth2-client/spec/deneb"
-	"github.com/attestantio/go-eth2-client/spec/phase0"
-	fastSsz "github.com/ferranbt/fastssz"
 )
 
 // Lengths of hashes and addresses in bytes.
@@ -517,187 +505,4 @@ func (s *HexBytes) UnmarshalJSON(input []byte) error {
 	}
 
 	return nil
-}
-
-// InclusionProof is a Merkle Multiproof of inclusion of a set of TransactionHashes
-type InclusionProof struct {
-	TransactionHashes  []Hash      `json:"transaction_hashes"`
-	GeneralizedIndexes []uint64    `json:"generalized_indexes"`
-	MerkleHashes       []*HexBytes `json:"merkle_hashes"`
-}
-
-// InclusionProofFromMultiProof converts a fastssz.Multiproof into an InclusionProof, without
-// filling the TransactionHashes
-func InclusionProofFromMultiProof(mp *fastSsz.Multiproof) *InclusionProof {
-	merkleHashes := make([]*HexBytes, len(mp.Hashes))
-	for i, h := range mp.Hashes {
-		merkleHashes[i] = new(HexBytes)
-		*(merkleHashes[i]) = h
-	}
-
-	leaves := make([]*HexBytes, len(mp.Leaves))
-	for i, h := range mp.Leaves {
-		leaves[i] = new(HexBytes)
-		*(leaves[i]) = h
-	}
-	generalIndexes := make([]uint64, len(mp.Indices))
-	for i, idx := range mp.Indices {
-		generalIndexes[i] = uint64(idx)
-	}
-	return &InclusionProof{
-		MerkleHashes:       merkleHashes,
-		GeneralizedIndexes: generalIndexes,
-	}
-}
-
-func (p *InclusionProof) String() string {
-	return JSONStringify(p)
-}
-
-// A wrapper struct over `builderSpec.VersionedSubmitBlockRequest`
-// to include constraint inclusion proofs
-type VersionedSubmitBlockRequestWithProofs struct {
-	Proofs *InclusionProof
-	*builderSpec.VersionedSubmitBlockRequest
-}
-
-// this is necessary, because the mev-boost-relay deserialization doesn't expect a "Version" and "Data" wrapper object
-// for deserialization. Instead, it tries to decode the object into the "Deneb" version first and if that fails, it tries
-// the "Capella" version. This is a workaround to make the deserialization work.
-//
-// NOTE(bolt): struct embedding of the VersionedSubmitBlockRequest is not possible for some reason because it causes the json
-// encoding to omit the `proofs` field. Embedding all of the fields directly does the job.
-func (v *VersionedSubmitBlockRequestWithProofs) MarshalJSON() ([]byte, error) {
-	switch v.Version {
-	case consensusSpec.DataVersionBellatrix:
-		return json.Marshal(struct {
-			Message          *v1.BidTrace                    `json:"message"`
-			ExecutionPayload *bellatrixSpec.ExecutionPayload `json:"execution_payload"`
-			Signature        phase0.BLSSignature             `json:"signature"`
-			Proofs           *InclusionProof                 `json:"proofs"`
-		}{
-			Message:          v.Bellatrix.Message,
-			ExecutionPayload: v.Bellatrix.ExecutionPayload,
-			Signature:        v.Bellatrix.Signature,
-			Proofs:           v.Proofs,
-		})
-	case consensusSpec.DataVersionCapella:
-		return json.Marshal(struct {
-			Message          *v1.BidTrace                  `json:"message"`
-			ExecutionPayload *capellaSpec.ExecutionPayload `json:"execution_payload"`
-			Signature        phase0.BLSSignature           `json:"signature"`
-			Proofs           *InclusionProof               `json:"proofs"`
-		}{
-			Message:          v.Capella.Message,
-			ExecutionPayload: v.Capella.ExecutionPayload,
-			Signature:        v.Capella.Signature,
-			Proofs:           v.Proofs,
-		})
-	case consensusSpec.DataVersionDeneb:
-		return json.Marshal(struct {
-			Message          *v1.BidTrace                `json:"message"`
-			ExecutionPayload *denebSpec.ExecutionPayload `json:"execution_payload"`
-			Signature        phase0.BLSSignature         `json:"signature"`
-			Proofs           *InclusionProof             `json:"proofs"`
-		}{
-			Message:          v.Deneb.Message,
-			ExecutionPayload: v.Deneb.ExecutionPayload,
-			Signature:        v.Deneb.Signature,
-			Proofs:           v.Proofs,
-		})
-	}
-
-	return nil, fmt.Errorf("unknown data version %d", v.Version)
-}
-
-func (v *VersionedSubmitBlockRequestWithProofs) String() string {
-	return JSONStringify(v)
-}
-
-// SignedConstraintsList are a list of proposer constraints that a builder must satisfy
-// in order to produce a valid bid. This is not defined on the
-// [spec](https://chainbound.github.io/bolt-docs/api/builder)
-// but it's useful as an helper type
-type SignedConstraintsList = []*SignedConstraints
-
-// Reference: https://chainbound.github.io/bolt-docs/api/builder
-type SignedConstraints struct {
-	Message   ConstraintsMessage  `json:"message"`
-	Signature phase0.BLSSignature `json:"signature"`
-}
-
-func (c *SignedConstraints) Digest() []byte {
-	hasher := sha256.New()
-	// NOTE: ignoring errors here
-	slotBytes := make([]byte, 8)
-	binary.LittleEndian.PutUint64(slotBytes, c.Message.Slot)
-
-	var top byte
-	if c.Message.Top {
-		top = 1
-	} else {
-		top = 0
-	}
-
-	hasher.Write(c.Message.Pubkey[:])
-	hasher.Write(slotBytes)
-	hasher.Write([]byte{top})
-
-	for _, tx := range c.Message.Transactions {
-		txHash := sha256.Sum256(*tx)
-		hasher.Write(txHash[:])
-	}
-
-	return hasher.Sum(nil)
-}
-
-func (c *SignedConstraints) VerifySignature(pubkey phase0.BLSPubKey) (bool, error) {
-	signingData := phase0.SigningData{ObjectRoot: phase0.Root(c.Digest()), Domain: phase0.Domain([]byte{109, 109, 111, 67})}
-	root, err := signingData.HashTreeRoot()
-	if err != nil {
-		return false, err
-	}
-
-	return bls.VerifySignatureBytes(root[:], c.Signature[:], pubkey[:])
-}
-
-// Reference: https://chainbound.github.io/bolt-docs/api/builder
-type ConstraintsMessage struct {
-	Pubkey       phase0.BLSPubKey `json:"pubkey"`
-	Slot         uint64           `json:"slot"`
-	Top          bool             `json:"top"`
-	Transactions []*HexBytes      `json:"transactions"`
-}
-
-// List of signed delegations
-type SignedDelegations = []*SignedDelegation
-
-type SignedDelegation struct {
-	Message   Delegation          `json:"message"`
-	Signature phase0.BLSSignature `json:"signature"`
-}
-
-func (d *SignedDelegation) Digest() []byte {
-	hasher := sha256.New()
-	// NOTE: ignoring errors here
-	hasher.Write(d.Message.ValidatorPubkey[:])
-	hasher.Write(d.Message.DelegateePubkey[:])
-	return hasher.Sum(nil)
-}
-
-// VerifySignature verifies the signature of a signed delegation. IMPORTANT: it uses the COMMIT_BOOST_DOMAIN
-// as the signing domain: [109, 109, 111, 67]
-func (d *SignedDelegation) VerifySignature(pubkey phase0.BLSPubKey) (bool, error) {
-	signingData := phase0.SigningData{ObjectRoot: phase0.Root(d.Digest()), Domain: phase0.Domain([]byte{109, 109, 111, 67})}
-	root, err := signingData.HashTreeRoot()
-	if err != nil {
-		return false, err
-	}
-
-	return bls.VerifySignatureBytes(root[:], d.Signature[:], pubkey[:])
-}
-
-type Delegation struct {
-	ValidatorPubkey phase0.BLSPubKey `json:"validator_pubkey"`
-	DelegateePubkey phase0.BLSPubKey `json:"delegatee_pubkey"`
 }
